requirements.txt
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
debugpy==1.8.17
watchdog==3.0.0